Merge in KI_ATLAS/atlas from feature/ATLAS-2136-postgresql-test-containers to master

* commit '08e34858ac90c84cfc77d601878e8e0f75eb8e07':
  ATLAS-2136: cleanup properties and h2 stuff
  ATLAS-2136: upgrade postgres to 17.1
  ATLAS-2136: Maven build running
  ATLAS-2136: integrate in export service point
  ATLAS-2136: Integrate with lidi
  ATLAS-2136: reset to current state
  ATLAS-2136: try to simplify dependencies
  ATLAS-2136: lets look at tekton build time
  ATLAS-2136: Use on bodi only for now
  ATLAS-2136: lets now try on tekton
  ATLAS-2136: lets see the tekton with config
  Try with label
  Use correct user id
  ATLAS-2136: try with runas-user
  ATLAS-2136: enable info logs
  ATLAS-2136: Enable test containers on tekton
  ATLAS-2136: Testcontainers first try

+++ location/src/test/java/ch/sbb/atlas/location/BaseLocationIntegrationTest.java
@@ -16,16 +16,17 @@ public abstract class BaseLocationIntegrationTest extends BaseControllerApiTest 
   }
 
   @AfterEach
-  void cleanUp(){
+  void cleanUp() {
     Arrays.stream(SloidType.values()).toList().forEach(sloidType -> {
       Set<String> allocatedSloid = sloidRepository.getAllocatedSloids(sloidType);
-      if(sloidType == SloidType.SERVICE_POINT){
-        sloidRepository.setAvailableSloidsToUnclaimed(allocatedSloid);
-        sloidRepository.deleteAllocatedSloids(allocatedSloid,SloidType.SERVICE_POINT);
-      }else {
-        sloidRepository.deleteAllocatedSloids(allocatedSloid,sloidType);
+      if (!allocatedSloid.isEmpty()) {
+        if (sloidType == SloidType.SERVICE_POINT) {
+          sloidRepository.setAvailableSloidsToUnclaimed(allocatedSloid);
+          sloidRepository.deleteAllocatedSloids(allocatedSloid, SloidType.SERVICE_POINT);
+        } else {
+          sloidRepository.deleteAllocatedSloids(allocatedSloid, sloidType);
+        }
       }
-
     });
   }
